# web/documentation/swagger.json
{"swagger":"2.0","info":{"title":"Valhalla API","description":"Valhalla API","version":""},"host":"localhost:5000","schemes":["http"],"consumes":["application/json","application/xml","application/gob","application/x-gob"],"produces":["application/json","application/xml","application/gob","application/x-gob"],"paths":{"/v1/godkjent":{"get":{"tags":["godkjent"],"summary":"vis godkjent","description":"vis godkjent","operationId":"godkjent#vis","produces":["application/standard_error+json","application/vnd.godkjentliste+json"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Godkjentliste"}},"404":{"description":"Not Found","schema":{"$ref":"#/definitions/Standard_error"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/Standard_error"}}},"schemes":["http"]},"post":{"tags":["godkjent"],"summary":"leggeTil godkjent","description":"Legge til en godjkent","operationId":"godkjent#leggeTil","produces":["application/standard_error+json","application/vnd.godkjent+json"],"parameters":[{"name":"payload","in":"body","description":"Legge til godkjent på listen","required":true,"schema":{"$ref":"#/definitions/LeggeTilGodkjentNyttelast"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Godkjent"}},"404":{"description":"Not Found","schema":{"$ref":"#/definitions/Standard_error"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/Standard_error"}}},"schemes":["http"]}},"/v1/helse":{"get":{"tags":["helse"],"summary":"vis helse","description":"OK hvis sunn","operationId":"helse#vis","produces":["application/standard_error+json","application/vnd.helse+json"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Helse"}},"404":{"description":"Not Found","schema":{"$ref":"#/definitions/Standard_error"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/Standard_error"}}},"schemes":["http"]}}},"definitions":{"Godkjent":{"title":"Mediatype identifier: application/vnd.godkjent+json; view=default","type":"object","properties":{"adresse":{"type":"string","description":"Ethereum adresse","example":"Cumque eius at et consequuntur maiores."},"navn":{"type":"string","description":"Navn på godkjent","example":"Consequuntur voluptas adipisci alias provident."}},"description":"Vis godjkent (default view)","example":{"adresse":"Cumque eius at et consequuntur maiores.","navn":"Consequuntur voluptas adipisci alias provident."},"required":["navn","adresse"]},"GodkjentCollection":{"title":"Mediatype identifier: application/vnd.godkjent+json; type=collection; view=default","type":"array","items":{"$ref":"#/definitions/Godkjent"},"description":"GodkjentCollection is the media type for an array of Godkjent (default view)","example":[{"adresse":"Cumque eius at et consequuntur maiores.","navn":"Consequuntur voluptas adipisci alias provident."},{"adresse":"Cumque eius at et consequuntur maiores.","navn":"Consequuntur voluptas adipisci alias provident."},{"adresse":"Cumque eius at et consequuntur maiores.","navn":"Consequuntur voluptas adipisci alias provident."}]},"Godkjentliste":{"title":"Mediatype identifier: application/vnd.godkjentliste+json; view=default","type":"object","properties":{"godkjentListe":{"$ref":"#/definitions/GodkjentCollection"}},"description":"Vis godkjent liste (default view)","example":{"godkjentListe":[{"adresse":"Cumque eius at et consequuntur maiores.","navn":"Consequuntur voluptas adipisci alias provident."},{"adresse":"Cumque eius at et consequuntur maiores.","navn":"Consequuntur voluptas adipisci alias provident."}]},"required":["godkjentListe"]},"Helse":{"title":"Mediatype identifier: application/vnd.helse+json; view=default","type":"object","properties":{"status":{"type":"string","description":"Status","example":"Eum non mollitia labore sapiente sint facilis."}},"description":"Helse (default view)","example":{"status":"Eum non mollitia labore sapiente sint facilis."},"required":["status"]},"LeggeTilGodkjentNyttelast":{"title":"LeggeTilGodkjentNyttelast","type":"object","properties":{"adresse":{"type":"string","description":"Ethereum adresse","example":"Cupiditate voluptatem fugit neque."},"navn":{"type":"string","description":"Navn på godkjent","example":"Officiis amet reiciendis assumenda in."}},"description":"Legge til godkjent på listen","example":{"adresse":"Cupiditate voluptatem fugit neque.","navn":"Officiis amet reiciendis assumenda in."},"required":["navn","adresse"]},"Standard_error":{"title":"Mediatype identifier: application/standard_error+json; view=default","type":"object","properties":{"code":{"type":"integer","description":"A code that describes the error","example":400,"format":"int64"},"message":{"type":"string","description":"A message that describes the error","example":"Bad Request"}},"description":"A standard error response (default view)","example":{"code":400,"message":"Bad Request"},"required":["code","message"]}}}